+++ libs/tktokenswap/conanfile.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2022 Cambridge Quantum Computing
+# Copyright 2019-2023 Cambridge Quantum Computing
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ class TktokenswapConan(ConanFile):
         "tklog/0.1.2@tket/stable",
         "tkassert/0.1.1@tket/stable",
         "tkrng/0.1.2@tket/stable",
-        "boost/1.80.0",
+        "boost/1.81.0",
     ]
 
     def config_options(self):